added test to methods

+++ src/main/java/com/example/springdocker/MyMathCalc.java
@@ -17,7 +17,7 @@ public class MyMathCalc {
     }
 
 
-    public float divide(float a, float b) {
+    public int divide(int a, int b) {
         return a/b;
     }
